refactor: :art: telegram funcs naming
